+++ tests/unit/Applet/Upload/appletUploadCommandOptionsMock.ts
@@ -1,0 +1,15 @@
+
+export const multiFileOptions = {
+	'applet-path': 'fakeAppletPath',
+	'entry-file-path': 'fakeEntryFilePath',
+	'organization-uid': 'fakeOrganizationUid',
+};
+
+export const singleFileOptions = {
+	'applet-path': 'fakeAppletPath',
+	'organization-uid': 'fakeOrganizationUid',
+};
+
+export const noUserOptions = {
+	'organization-uid': 'fakeOrganizationUid',
+};