+++ tests/unit/Applet/Upload/appletUploadFacadeMock.ts
@@ -1,0 +1,8 @@
+import * as sinon from 'sinon';
+
+export default {
+	updateSingleFileApplet: sinon.fake(),
+	updateMultiFileApplet: sinon.fake(),
+	createSingleFileApplet: sinon.fake(),
+	createMultiFileFileApplet: sinon.fake(),
+};