+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.5.0] - 2019-11-28
 ### Added
 - Support multiple files of applet in Webpack Plugin
+- Upload multi file applet
 
 ### Fixed
 - Allow CORS in webpack plugin 8090 emulator proxy port for develop applet externally